v0.2
Very different, still not finished at all.
Progress: I’ve given up C, the hex-files are always too large for the
ATTiny13A. The most recent files are in /tree. The code is now in
Assembly and manages to fade one LED in and out. If I now get it to
also fade other LEDs out of phase, that would be it.